+++ examples/alice_dio/lib/main.dart
@@ -20,7 +20,7 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   late final AliceDioAdapter _aliceDioAdapter = AliceDioAdapter();
 
-  final configuration = AliceConfiguration();
+  final configuration = AliceConfiguration(showShareButton: true);
   late final Alice _alice = Alice(configuration: configuration)
     ..addAdapter(_aliceDioAdapter);