+++ packages/alice/CHANGELOG.md
@@ -1,4 +1,10 @@
-## 1.1.1
+## Unreleased
+
+* Added support for share curl option (by @thanhnt-mdev).
+* Added sharePositionOrigin support for iOS (by @thanhnt-mdev).
+* Added FAB menu for share and share curl.
+
+## 1.1.2
 * Bumped dependencies.
 
 # 1.1.0